+++ wow_dbc/src/vanilla_tables/lock.rs
@@ -198,7 +198,7 @@ impl TryFrom<i32> for Type {
 }
 
 impl Type {
-    const fn as_int(&self) -> i32 {
+    pub const fn as_int(&self) -> i32 {
         match self {
             Self::None => 0,
             Self::ItemRequired => 1,